Fixes sensi parameters deps cycles
